Adds schedule_nightly_graph so git information is available to nightly .taskcluster.yml
